requirements.txt
psycopg==3.2.3
psycopg-binary==3.2.3
typing_extensions==4.12.2
python-dotenv==1.0.1
